% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{achievement}
\alias{achievement}
\title{Teacher Expectancies on Student Achievement}
\format{
A dataframe with 40 rows and four variabels
\describe{
\item{exp}{Teacher expectancy based on an intelligence test given to the student the previous year.}
\item{soc}{The average observer rating of social warmth.}
\item{inp}{The average observer rating of input to the student.}
\item{ach}{Studen achievement measured using a score in the test at the end of the semester.}
}
}
\usage{
data(achievement)
}
\description{
Teacher Expectancies on Student Achievement
}
\examples{
data(achievement)
}
\keyword{datasets}
