% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{alienation}
\alias{alienation}
\title{Stability of Alienation}
\format{
A \eqn{6 \times 6} covariance matrix of the following variables
\describe{
\item{anomie67}{Anomie measured in 1967.}
\item{powerless67}{Powerlessness measured in 1967.}
\item{anomie71}{Anomie measured in 1971.}
\item{powerless71}{Powerlessness measured in 1971.}
\item{edu}{Education.}
\item{sei}{Duncan's occupation status index (SEI).}
}
}
\usage{
data(alienation)
}
\description{
Stability of Alienation
}
\examples{
data(alienation)
}
\keyword{datasets}
