% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_lav}
\alias{boot_lav}
\title{Bootstrap SEM Parameter Estimates}
\usage{
boot_lav(
  data,
  model,
  raw_data = TRUE,
  mean_vector = NULL,
  n = NULL,
  B = 2000,
  cluster = TRUE,
  cores = NULL,
  nb = FALSE,
  ...
)
}
\arguments{
\item{data}{Matrix of sample data to bootstrap.}

\item{model}{Specified model to fit following \code{lavaan} notation.}

\item{raw_data}{Logical.
If \code{TRUE}, \code{data} is a sample raw data.
If \code{FALSE}, \code{data} is a sample variance-covariance matrix.}

\item{mean_vector}{A sample mean vector.
Ignored if \code{raw_data = TRUE}.}

\item{n}{Sample size.
Ignored if \code{raw_data = TRUE}.}

\item{B}{Number of bootstrap resamples.}

\item{cluster}{Logical.
If \code{TRUE}, parallelize computations using a cluster.}

\item{cores}{Integer.
Number of cores to use.
Defaults to total number of threads minus 1.}

\item{nb}{Logical.
If \code{TRUE}, performs nonparametric bootstrapping.
Ignored if \code{raw_data = FALSE}.}

\item{...}{Arguments to pass to \code{fitFUN}.}
}
\description{
Perform nonparametric and parametric bootstrapping
with percentile and bias corrected confidence intervals
on SEM parameter estimates.
}
\seealso{
Other bootstrap functions: 
\code{\link{boot_fit}()},
\code{\link{boot_nb}()},
\code{\link{boot_pb_med_simple}()},
\code{\link{boot_pb_vm}()},
\code{\link{boot_pb}()}
}
\concept{bootstrap functions}
\keyword{bootstraping}
