% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_nb}
\alias{boot_nb}
\title{Nonparametric Bootstrap}
\usage{
boot_nb(data, B = 2000)
}
\arguments{
\item{data}{Matrix of sample data to bootstrap.}

\item{B}{Number of bootstrap resamples.}
}
\value{
Returns a list of nonparametric bootstrap resamples.
}
\description{
Generates \code{B} number of nonparametric bootstrap
resamples from the original sample data.
}
\examples{
B <- 5
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
mu <- c(100, 100, 100)
data <- dat_mvn(
  n = 100,
  Sigma = Sigma,
  mu = mu
)
boot_nb_resamples <- boot_nb(data = data, B = B)
}
\seealso{
Other bootstrap functions: 
\code{\link{boot_fit}()},
\code{\link{boot_pb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstraping}
