% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_pb}
\alias{boot_pb}
\title{Parametric Bootstrap}
\usage{
boot_pb(n, Sigma, mu, B = 2000)
}
\arguments{
\item{n}{Sample size.}

\item{Sigma}{Estimated variance-covariance matrix
from the original sample data.}

\item{mu}{Estimated mean vector from the original sample data.}

\item{B}{Number of bootstrap resamples.}
}
\value{
Returns a list of parametric bootstrap resamples.
}
\description{
Generates \code{B} number of parametric bootstrap
resamples from the original sample data.
Data is generated from a multivariate normal distribution
using the estimated variance-covariance matrix and mean vector.
}
\examples{
B <- 5
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
mu <- c(100, 100, 100)
data <- dat_mvn(
  n = 100,
  Sigma = Sigma,
  mu = mu
)
n <- nrow(data)
est_Sigma <- cov(data)
est_mu <- colMeans(data)
boot_pb_resamples <- boot_pb(n = n, Sigma = est_Sigma, mu = est_mu, B = B)
}
\seealso{
Other bootstrap functions: 
\code{\link{boot_fit}()},
\code{\link{boot_nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstraping}
