% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_pb_med_simple}
\alias{boot_pb_med_simple}
\title{Parametric Bootstrap (Simple Mediation Model)}
\usage{
boot_pb_med_simple(
  n,
  alpha,
  tau_prime,
  beta,
  delta_M,
  delta_Y,
  sigma2_epsilon_M,
  sigma2_epsilon_Y,
  B = 2000,
  rFUN_X,
  X_args
)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Path from \eqn{X} to \eqn{M} (\eqn{\alpha}).}

\item{tau_prime}{Path from \eqn{X} to \eqn{Y} (\eqn{\tau^{\prime}}).}

\item{beta}{Path from \eqn{M} to \eqn{Y} (\eqn{\beta}).}

\item{delta_M}{Intercept for the first equation (\eqn{\delta_M}).}

\item{delta_Y}{Intercept for the second equation (\eqn{\delta_Y}).}

\item{sigma2_epsilon_M}{Variance of \eqn{\epsilon_M} (\eqn{\sigma^{2}_{\epsilon_{M}}}).}

\item{sigma2_epsilon_Y}{Variance of \eqn{\epsilon_Y} (\eqn{\sigma^{2}_{\epsilon_{Y}}}).}

\item{B}{Number of bootstrap resamples.}

\item{rFUN_X}{The distribution function
used to generate values of \eqn{\mathbf{X}}.
The default value is \code{rnorm}
for the Gaussian probability density function.}

\item{X_args}{List of arguments
to pass to \code{rFUN_X}.}
}
\value{
Returns a list of parametric bootstrap resamples.
}
\description{
Generates \code{B} number of parametric bootstrap
resamples from the original sample data generated
fitted with the simple mediation model,
\eqn{M_i = \delta_M + \alpha X_i + \epsilon_{M_i}} and
\eqn{Y_i = \delta_Y + \tau^{\prime} X_i + \beta M_i + \epsilon_{Y_i}}.
The user can specify the distribution and parameters of \eqn{X}.
Variables \eqn{M} and \eqn{Y} are generated
using values of variable \eqn{X}
and parameters provided assuming that the residuals
are normally distributed
with a mean of 0 and a given variance.
}
\examples{
n <- 100
alpha <- 0.26^(1 / 2)
tau_prime <- 0
beta <- 0.26^(1 / 2)
delta_M <- delta_Y <- 0
sigma2_epsilon_M <- 1 - alpha^2
sigma2_epsilon_Y <- 1 - beta^2 - tau_prime^2 - 2 * alpha * beta * tau_prime
B <- 5
rFUN_X <- rnorm
X_args <- list(mean = 0, sd = 1)
}
\seealso{
Other bootstrap functions: 
\code{\link{boot_fit}()},
\code{\link{boot_lav}()},
\code{\link{boot_nb}()},
\code{\link{boot_pb_vm}()},
\code{\link{boot_pb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstraping}
