% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_pb_vm}
\alias{boot_pb_vm}
\title{Parametric Bootstrap VM}
\usage{
boot_pb_vm(n, Sigma_dot, skew, kurt, rescale = TRUE, mu, sigma2, B = 2000)
}
\arguments{
\item{n}{Sample size.}

\item{Sigma_dot}{\eqn{p \times p} correlation matrix (\eqn{\dot{\Sigma}}).}

\item{skew}{Skewness vector.
If a single value is given, the function assumes that
the \eqn{p} elements in the vector will have the same skewness.}

\item{kurt}{Kurtosis vector.
If a single value is given, the function assumes that
the \eqn{p} elements in the vector will have the same kurtosis.}

\item{rescale}{Logical.
Rescale the data using means \code{mu} (\eqn{\mu})
and variances \code{sigma2} (\eqn{\sigma^2}) provided.
Note that the actual means and variances that the final data set will have
are drawn randomly from a normal distrbution
using \code{mu} and \code{sigma2} as parameters.}

\item{mu}{Vector of means
corresponding to each element of \code{Sigma_dot} (\eqn{\dot{\Sigma}})
that will be used as population means when rescale is \code{TRUE}.
If a single value is given, the function assumes that
the \eqn{p} elements in the vector will have the same \code{mu}.}

\item{sigma2}{Vector of variances
corresponding to each element of \code{Sigma_dot} (\eqn{\dot{\Sigma}})
that will be used as population variances when rescale is \code{TRUE}.
If a single value is given, the function assumes that
the \eqn{p} elements in the vector will have the same \code{sigma2}.}

\item{B}{Number of bootstrap resamples.}
}
\value{
Returns a list of parametric bootstrap resamples.
}
\description{
Generates \code{B} number of parametric bootstrap
resamples from the original sample data
using \code{gendat_vm}.
}
\seealso{
Other bootstrap functions: 
\code{\link{boot_fit}()},
\code{\link{boot_lav}()},
\code{\link{boot_nb}()},
\code{\link{boot_pb_med_simple}()},
\code{\link{boot_pb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstraping}
