% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_misc.R
\name{ci_hit}
\alias{ci_hit}
\title{CI hit}
\usage{
ci_hit(ll, parameter, ul)
}
\arguments{
\item{ll}{Lower limit.}

\item{parameter}{Population parameter.}

\item{ul}{Upper limit.}
}
\value{
Returns \code{TRUE} or \code{FALSE} based on the test.
}
\description{
Tests if the confidence intervals
contain the population parameter.
}
\seealso{
Other confidence interval miscellaneous functions: 
\code{\link{ci_shape}()},
\code{\link{ci_width}()},
\code{\link{zero_hit}()}
}
\concept{confidence interval miscellaneous functions}
\keyword{ci}
