% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_misc.R
\name{ci_shape}
\alias{ci_shape}
\title{CI shape}
\usage{
ci_shape(ll, est, ul)
}
\arguments{
\item{ll}{Lower limit.}

\item{est}{Parameter estimate.}

\item{ul}{Upper limit.}
}
\value{
Returns confidence interval shape.
}
\description{
Calculates shape of the confidence intervals.
}
\seealso{
Other confidence interval miscellaneous functions: 
\code{\link{ci_hit}()},
\code{\link{ci_width}()},
\code{\link{zero_hit}()}
}
\concept{confidence interval miscellaneous functions}
\keyword{ci}
