% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_misc.R
\name{ci_wald}
\alias{ci_wald}
\title{Wald Confidence Intervals}
\usage{
ci_wald(est, se)
}
\arguments{
\item{est}{Parameter estimate.}

\item{se}{Standard error.}
}
\value{
Returns a vector with the following elements:
\describe{
\item{se}{Standard error.}
\item{z}{\code{z} statistic.}
\item{p}{\code{p}-value.}
\item{sig_001}{Logical. Tests if \code{p} < 0.001.}
\item{sig_01}{Logical. Tests if \code{p} < 0.01.}
\item{sig_05}{Logical. Tests if \code{p} < 0.05.}
\item{ll_001}{Lower limit. Alpha 0.001.}
\item{ll_01}{Lower limit. Alpha 0.01.}
\item{ll_05}{Lower limit. Alpha 0.05.}
\item{ul_05}{Upper limit. Alpha 0.05.}
\item{ul_01}{Upper limit. Alpha 0.01.}
\item{ul_001}{Upper limit. Alpha 0.001.}
\item{zero_hit_001}{Logical. Tests if \code{ll_001} < 0 < \code{ul_001}.}
\item{zero_hit_01}{Logical. Tests if \code{ll_01} < 0 < \code{ul_01}.}
\item{zero_hit_05}{Logical. Tests if \code{ll_05} < 0 < \code{ul_05}.}
\item{ci_width_001}{\code{ll_001} - \code{ul_001}.}
\item{ci_width_01}{\code{ll_01} - \code{ul_01}.}
\item{ci_width_05}{\code{ll_05} - \code{ul_05}.}
\item{ci_shape_001}{Confidence interval shape. Alpha 0.001.}
\item{ci_shape_01}{Confidence interval shape. Alpha 0.01.}
\item{ci_shape_05}{Confidence interval shape. Alpha 0.05.}
}
Note that \code{se}, \code{z}, \code{p}, and \code{sig}
are based on the normal theory.
Confidence limits for \code{\link{ci_wald}} are based on the normal theory.
Ajustments for asymmetry for \code{\link{ci_bc}}
and \code{\link{ci_bca}} are made in the confidence limits.
}
\description{
Generates symmetric confidence intervals using the Wald method.
}
\examples{
ci_wald(est = 0.26^(1 / 2), se = 0.25)
}
\seealso{
Other confidence interval functions: 
\code{\link{ci_bca}()},
\code{\link{ci_bc}()},
\code{\link{ci_quantile}()}
}
\concept{confidence interval functions}
\keyword{ci}
