% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_misc.R
\name{ci_width}
\alias{ci_width}
\title{CI width}
\usage{
ci_width(ll, ul)
}
\arguments{
\item{ll}{Lower limit.}

\item{ul}{Upper limit.}
}
\value{
Returns confidence interval width
}
\description{
Calculates width of the confidence intervals.
}
\seealso{
Other confidence interval miscellaneous functions: 
\code{\link{ci_hit}()},
\code{\link{ci_shape}()},
\code{\link{zero_hit}()}
}
\concept{confidence interval miscellaneous functions}
\keyword{ci}
