% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{dat_linreg}
\alias{dat_linreg}
\title{Generate Random Data From a Linear Regression Model.}
\usage{
dat_linreg(n, beta, rFUN_X = rnorm, rFUN_y = rnorm, X_args, y_args)
}
\arguments{
\item{n}{Positive integer.
Sample size.}

\item{beta}{Numeric vector.
\eqn{k \times 1} vector of \eqn{k} regression parameters.}

\item{rFUN_X}{Function.
The distribution function
used to generate values of \eqn{\mathbf{X}}.
The default value is \code{\link[stats:rnorm]{stats::rnorm()}}
for the Gaussian probability density function.}

\item{rFUN_y}{Function.
The distribution function
used to generate values of the residuals \eqn{\epsilon}.
The default value is \code{\link[stats:rnorm]{stats::rnorm()}}
for the Gaussian probability density function.}

\item{X_args}{Named list.
List of arguments
to pass to \code{rFUN_X}.}

\item{y_args}{Named list.
List of arguments
to pass to \code{rFUN_y}.}
}
\value{
Returns a list with two elements \code{X} and \code{y}.
\itemize{
\item \code{X} is the data matrix,
that is an \eqn{n \times k}  matrix
of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.
\item \code{y} \eqn{n \times 1} vector of observations on the regressand.
}
}
\description{
Generates data from a \eqn{k}-variable linear regression model.
}
\details{
Randomly generates the data matrix
\eqn{\mathbf{X}}
and the regressand data
\eqn{\mathbf{y}}
using specified population parameters defined by
\eqn{\mathbf{y}_{n \times 1}
  =
  \mathbf{X}_{n \times k}
  \boldsymbol{\beta}_{k \times 1} +
  \boldsymbol{\epsilon}_{n \times 1}}.
Refer to \code{\link[=dat_linreg_X]{dat_linreg_X()}}
on how \eqn{\mathbf{X}} is generated
and \code{\link[=dat_linreg_y]{dat_linreg_y()}}
on how \eqn{\mathbf{y}} is generated.
}
\examples{
data <- dat_linreg(
  n = 100,
  beta = c(.5, .5, .5),
  rFUN_X = rnorm,
  rFUN_y = rnorm,
  X_args = list(mean = 0, sd = 1),
  y_args = list(mean = 0, sd = 1)
)
}
\seealso{
Other data generating functions: 
\code{\link{dat_linreg_X}()},
\code{\link{dat_linreg_y}()},
\code{\link{dat_mvn_a}()},
\code{\link{dat_mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{data generating functions}
\keyword{data}
