% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{dat_linreg_X}
\alias{dat_linreg_X}
\title{Generate Data Matrix (X) From a Linear Regression Model.}
\usage{
dat_linreg_X(n, k, constant = TRUE, rFUN_X = rnorm, ...)
}
\arguments{
\item{n}{Positive integer.
Sample size.}

\item{k}{Positive integer.
Number of regressors.}

\item{constant}{Logical.
An option to include or to exclude the vector of constants.
If \code{TRUE}, the vector of constants is included.
If \code{FALSE}, the vector of constants is excluded.}

\item{rFUN_X}{Function.
The distribution function
used to generate values of \eqn{\mathbf{X}}.
The default value is \code{\link[stats:rnorm]{stats::rnorm()}}
for the Gaussian probability density function.}

\item{...}{Arguments to pass to \code{rFUN_X}.}
}
\value{
If \code{constant = TRUE},
returns an \eqn{n \times k} numeric matrix
where the first column consists of 1s.
If \code{constant = FALSE},
returns an \eqn{n \times k - 1} numeric matrix.
}
\description{
Generates random data matrix from a \eqn{k}-variable linear regression model.
}
\details{
Randomly generates the data matrix
(\eqn{\mathbf{X}}),
that is an
\eqn{n \times k}
dimensional matrix of
\eqn{n}
observations of
\eqn{k}
regressors,
which includes a regressor whose value is 1 for each observation.
The data generating function is supplied by the argument
\code{rFUN_X}
(\code{\link[stats:rnorm]{stats::rnorm()}} is the default value).
Additional arguments to \code{rFUN_X} are supplied using the
\code{...} argument.
The data matix is also called the design matrix and model matrix.
}
\examples{
X <- dat_linreg_X(
  n = 100,
  k = 3,
  constant = TRUE,
  rFUN_X = rnorm,
  mean = 100,
  sd = 15
)
}
\seealso{
Other data generating functions: 
\code{\link{dat_linreg_y}()},
\code{\link{dat_linreg}()},
\code{\link{dat_mvn_a}()},
\code{\link{dat_mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{data generating functions}
\keyword{data}
