% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{dat_linreg_y}
\alias{dat_linreg_y}
\title{Generate Regressand Data (y) From a Linear Regression Model.}
\usage{
dat_linreg_y(X, beta, rFUN_y = rnorm, ...)
}
\arguments{
\item{X}{Matrix.
The data matrix, that is an \eqn{n \times k}  matrix
of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.
Also called the design matrix and model matrix.}

\item{beta}{Numeric vector.
\eqn{k \times 1} vector of \eqn{k} regression parameters.}

\item{rFUN_y}{Function.
The distribution function
used to generate values of the residuals \eqn{\epsilon}.
The default value is \code{\link[stats:rnorm]{stats::rnorm()}}
for the Gaussian probability density function.}

\item{...}{Arguments to pass to \code{rFUN_y}.}
}
\description{
Generates regressand data from a \eqn{k}-variable linear regression model.
}
\details{
Randomly generates the regressand data (\eqn{\mathbf{y}})
using specified population parameters defined by
\eqn{\mathbf{y}_{n \times 1}
  =
  \mathbf{X}_{n \times k}
  \boldsymbol{\beta}_{k \times 1} +
  \boldsymbol{\epsilon}_{n \times 1}
}.
The distribution of \eqn{\epsilon} is supplied by the argument
\code{rFUN_y}
(\code{\link[stats:rnorm]{stats::rnorm()}} is the default value).
Additional arguments to \code{rFUN_y} are supplied using the
\code{...} argument.
By default,
\eqn{\epsilon} is assumed to be normally distributed
with a mean of 0 and a variance of 1
(\eqn{
  \mathcal{N}
  \sim
  \left(
    \mu_{\epsilon} = 0,
    \sigma_{\epsilon}^{2} = 1
  \right)
}).
}
\examples{
X <- dat_linreg_X(
  n = 100,
  k = 3,
  constant = TRUE,
  rFUN_X = rnorm,
  mean = 0,
  sd = 1
)
y <- dat_linreg_y(
  X = X,
  beta = c(.5, .5, .5),
  rFUN_y = rnorm,
  mean = 0,
  sd = 1
)
}
\seealso{
Other data generating functions: 
\code{\link{dat_linreg_X}()},
\code{\link{dat_linreg}()},
\code{\link{dat_mvn_a}()},
\code{\link{dat_mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{data generating functions}
\keyword{data}
