% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_med_serial2}
\alias{dat_med_serial2}
\title{Serial Mediation Model Data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 123 rows and 4 columns.
}
\usage{
data(dat_med_serial2)
}
\description{
Serial mediation model data
from a study of presumed media influence.
}
\examples{
data(dat_med_serial2)
}
\keyword{datasets}
