% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_med_simple_lat}
\alias{dat_med_simple_lat}
\title{Simple Mediation Model Using Latent Variables Data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 9 columns.
}
\usage{
data(dat_med_simple_lat)
}
\description{
Simple mediation model data using latent variables.
}
\examples{
data(dat_med_simple_lat)
}
\keyword{datasets}
