% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{dat_mvn}
\alias{dat_mvn}
\title{Generate Multivariate Normal Data.}
\usage{
dat_mvn(n, Sigma, mu = NULL, ...)
}
\arguments{
\item{n}{Sample size.}

\item{Sigma}{\eqn{p \times p} variance-covariance matrix.}

\item{mu}{\eqn{p} dimensional mean vector. Defaults to zeros if unspecified.}

\item{...}{Arguments that can be passed to \code{\link[MASS:mvrnorm]{MASS::mvrnorm}}.}
}
\value{
Returns an \eqn{n \times p} multivariate normal data matrix generated
using the variance-covariance matrix
and the mean vector provided.
}
\description{
Generates multivariate normal data from
a \eqn{p \times p} variance-covariance matrix and
\eqn{p} dimensional mean vector.
}
\details{
Data is generated from a multivariate normal distrubution
given by
\eqn{
    \mathcal{N}
    \sim
    \left(
      \mathbf{\mu_{p \times 1}}, \mathbf{\Sigma_{p \times p}}
    \right)
   }
where
\eqn{\mathcal{N}}
has the density function
\eqn{
  \frac{
    \exp
    \left[ - \frac{1}{2} \left( \mathbf{X} - \boldsymbol{\mu} \right)^{T} \right]
    \boldsymbol{\Sigma}^{-1} \left( \mathbf{X} - \boldsymbol{\mu} \right)
  }
  {
  \sqrt{ \left( 2 \pi \right)^{k} | \boldsymbol{\Sigma} | }
  }
}.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
mu <- c(100, 100, 100)
data <- dat_mvn(
  n = 100,
  Sigma = Sigma,
  mu = mu
)
}
\seealso{
Other data generating functions: 
\code{\link{dat_linreg_X}()},
\code{\link{dat_linreg_y}()},
\code{\link{dat_linreg}()},
\code{\link{dat_mvn_a}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{data generating functions}
\keyword{data}
