% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{dat_mvn_a}
\alias{dat_mvn_a}
\title{Generate Multivariate Normal Data from the A Matrix and Variances of Observed Variables}
\usage{
dat_mvn_a(n, A, sigma2, F, I, mu = NULL, ...)
}
\arguments{
\item{n}{Sample size.}

\item{A}{Asymmetric paths,
such as regression coefficients and factor loadings.}

\item{sigma2}{Vector of variances (\eqn{\sigma^2}).}

\item{F}{Filter matrix
used to select the observed variables.}

\item{I}{Identity matrix.}

\item{mu}{\eqn{p} dimensional mean vector. Defaults to zeros if unspecified.}

\item{...}{Arguments that can be passed to \code{\link[MASS:mvrnorm]{MASS::mvrnorm}}.}
}
\value{
Returns an \eqn{n \times p} multivariate normal data matrix generated
using the variance-covariance matrix derived from the RAM matrices
and the mean vector provided.
}
\description{
Generates multivariate normal data from
a \eqn{p \times p} A matrix,
\eqn{p} dimensional vector of variances of observed variables, and
\eqn{p} dimensional vector of means.
}
\details{
The function interally uses the
\code{\link{ram_s}}
function
to derive the
\eqn{\Sigma_{p \times p}}
matrix from the matices provided.
The generated
\eqn{\Sigma_{p \times p}}
matrix is then used together with the
\eqn{p} dimensional
vector of means to generate data using
\code{\link{dat_mvn}}.
}
\examples{
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
sigma2 <- c(15^2, 15^2, 15^2)
F <- I <- diag(3)
mu <- c(100, 100, 100)
dat <- dat_mvn_a(
  n = 1000,
  A = A,
  sigma2 = sigma2,
  F = F,
  I = I,
  mu = mu
)
}
\seealso{
Other data generating functions: 
\code{\link{dat_linreg_X}()},
\code{\link{dat_linreg_y}()},
\code{\link{dat_linreg}()},
\code{\link{dat_mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{data generating functions}
\keyword{data}
