% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{dat_mvn_ram}
\alias{dat_mvn_ram}
\title{Generate Multivariate Normal Data from RAM Matrices}
\usage{
dat_mvn_ram(n, A, S, F, I, mu = NULL, ...)
}
\arguments{
\item{n}{Sample size.}

\item{A}{Asymmetric paths,
such as regression coefficients and factor loadings.}

\item{S}{Symmetric matrix
representing variances and covariances.}

\item{F}{Filter matrix
used to select the observed variables.}

\item{I}{Identity matrix.}

\item{mu}{\eqn{p} dimensional mean vector. Defaults to zeros if unspecified.}

\item{...}{Arguments that can be passed to \code{\link[MASS:mvrnorm]{MASS::mvrnorm}}.}
}
\value{
Returns an \eqn{n \times p} multivariate normal data matrix generated
using the variance-covariance matrix derived from the RAM matrices
and the mean vector provided.
}
\description{
Generates multivariate normal data from
the RAM matices, and
\eqn{p} dimensional vector of means.
}
\details{
The function interally uses the
\code{\link{ram}}
function
to derive the
\eqn{\Sigma_{p \times p}}
matrix from the matices provided.
The generated
\eqn{\Sigma_{p \times p}}
matrix is then used together with the
\eqn{p} dimensional
vector of means to generate data using
\code{\link{dat_mvn}}.
}
\examples{
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
S <- F <- I <- diag(3)
S[1, 1] <- 225
S[2, 2] <- 166.5
S[3, 3] <- 166.5
mu <- c(100, 100, 100)
data <- dat_mvn_ram(
  n = 100,
  A = A,
  S = S,
  F = F,
  I = I,
  mu = mu
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{data}
