% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{gd_delta}
\alias{gd_delta}
\title{Delta}
\usage{
gd_delta(w, X, y, aFUN, ...)
}
\arguments{
\item{w}{Vector of weights.}

\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the dependent variable.}

\item{aFUN}{Activation function.}

\item{...}{Arguments to pass to \code{aFUN}.}
}
\value{
Returns delta.
}
\description{
\eqn{X^{\prime} \mathbf{\hat{y}} - \mathbf{y}}
}
\keyword{gd}
