% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{gd_z}
\alias{gd_z}
\title{Predictions}
\usage{
gd_z(w, X)
}
\arguments{
\item{w}{Vector of weights.}

\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}
}
\value{
Returns model predictions.
}
\description{
\eqn{\mathbf{z} = \mathbf{X} \mathbf{w}}
}
\keyword{gd}
