% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat.R
\name{gendat}
\alias{gendat}
\title{Generate Random Data.}
\usage{
gendat(n, FUN = gendat_mvn, ...)
}
\arguments{
\item{n}{Sample size.}

\item{FUN}{Function to use to specify the model.
The default function is \code{\link{gendat_mvn}}
which generates data from a multivariate normal distribution
defined by a \eqn{\mu} vector (\code{mu}) (which defaults to zero if unspecified)
and a \eqn{\Sigma} matrix of covariances (\code{Sigma}).}

\item{...}{Arguments to pass to \code{FUN}.}
}
\description{
Generates data from a particular statistical model.
}
\details{
The default function is \code{\link{gendat_mvn}}
which generates multivariate normal data from a
\eqn{p \times p}
variance-covariance matrix and
\eqn{p} dimensional
vector of means.
}
\examples{
# Multivariate normal data
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
mu <- c(100, 100, 100)
mvn <- gendat(
  n = 100,
  FUN = gendat_mvn,
  Sigma = Sigma,
  mu = mu
)
# Multivariate normal data using RAM
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
S <- F <- I <- diag(3)
S[1, 1] <- 225
S[2, 2] <- 166.5
S[3, 3] <- 166.5
mu <- c(100, 100, 100)
mvn_ram <- gendat(
  n = 100,
  FUN = gendat_mvn_ram,
  A = A,
  S = S,
  F = F,
  I = I,
  mu = mu
)
# Multivariate normal data using RAM and sigma2
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
sigma2 <- c(15^2, 15^2, 15^2)
F <- I <- diag(3)
mu <- c(100, 100, 100)
mvn_a <- gendat(
  n = 100,
  FUN = gendat_mvn_a,
  A = A,
  sigma2 = sigma2,
  F = F,
  I = I,
  mu = mu
)
# Linear regression
linreg <- gendat(
  n = 100,
  FUN = gendat_linreg,
  beta = c(.5, .5, .5),
  rFUN_X = rnorm,
  rFUN_y = rnorm,
  X_args = list(mean = 0, sd = 1),
  y_args = list(mean = 0, sd = 1)
)
# Logistic regression
logreg <- gendat(
  n = 100,
  FUN = gendat_logreg,
  beta = c(.5, .5, .5),
  rFUN_X = rnorm,
  mean = 0,
  sd = 1
)
}
\seealso{
Other data generating functions: 
\code{\link{gendat_linreg_X}()},
\code{\link{gendat_linreg_y}()},
\code{\link{gendat_linreg}()},
\code{\link{gendat_mvn_a}()},
\code{\link{gendat_mvn_fe}()},
\code{\link{gendat_mvn}()},
\code{\link{gendat_vm}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{data generating functions}
\keyword{data}
