% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat.R
\name{gendat_logreg}
\alias{gendat_logreg}
\title{Generate Data Based on the Logistic Regression Model.}
\usage{
gendat_logreg(n, beta, rFUN_X = rnorm, ...)
}
\arguments{
\item{n}{Sample size.}

\item{beta}{\eqn{k \times 1} vector of \eqn{k} regression parameters.}

\item{rFUN_X}{The distribution function
used to generate values of \eqn{\mathbf{X}}.
The default value is \code{rnorm}
for the Gaussian probability density function.}

\item{...}{Arguments to pass to \code{rFUN_X}.}
}
\value{
Returns a list with two elements \code{X} and \code{y}.
\itemize{
\item \code{X} is the data matrix,
that is an \eqn{n \times k}  matrix
of \eqn{n} observations of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.
\item \code{y} \eqn{n \times 1} vector of observations on the regressand.
}
}
\description{
Generates data based on the logistic regression model
using the inverse logit link function.
}
\details{
Randomly generates the data matrix
\eqn{\mathbf{X}}
and the regressand data
\eqn{\mathbf{y}}.
\eqn{\mathbf{y}_{n \times 1}} is generated
from a binomial distribution
where
\eqn{p \left( x \right)}
is equal to
\eqn{
    \frac{
      1
    }
    {
      1 - e^{- \left( \mathbf{X}_{n \times k} \boldsymbol{\beta}_{n \times 1} \right)}
    }
  }.
}
\examples{
data <- gendat_logreg(
  n = 100,
  beta = c(.5, .5, .5),
  rFUN_X = rnorm,
  mean = 0,
  sd = 1
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{data}
