% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat.R
\name{gendat_med_simple}
\alias{gendat_med_simple}
\title{Generate Simple Mediation Model Data}
\usage{
gendat_med_simple(
  n,
  alpha,
  tau_prime,
  beta,
  delta_M,
  delta_Y,
  sigma2_epsilon_M,
  sigma2_epsilon_Y,
  rFUN_X,
  ...
)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Path from \eqn{X} to \eqn{M} (\eqn{\alpha}).}

\item{tau_prime}{Path from \eqn{X} to \eqn{Y} (\eqn{\tau^{\prime}}).}

\item{beta}{Path from \eqn{M} to \eqn{Y} (\eqn{\beta}).}

\item{delta_M}{Intercept for the first equation (\eqn{\delta_M}).}

\item{delta_Y}{Intercept for the second equation (\eqn{\delta_Y}).}

\item{sigma2_epsilon_M}{Variance of \eqn{\epsilon_M} (\eqn{\sigma^{2}_{\epsilon_{M}}}).}

\item{sigma2_epsilon_Y}{Variance of \eqn{\epsilon_Y} (\eqn{\sigma^{2}_{\epsilon_{Y}}}).}

\item{rFUN_X}{The distribution function
used to generate values of \eqn{\mathbf{X}}.
The default value is \code{rnorm}
for the Gaussian probability density function.}

\item{...}{Arguments to pass to \code{rFUN_X}.}
}
\description{
Generates data from a simple mediation model.
}
\details{
The simple mediation model is defined by
\eqn{M_i = \delta_M + \alpha X_i + \epsilon_{M_i}}, and
\eqn{Y_i = \delta_Y + \tau^{\prime} X_i + \beta M_i + \epsilon_{Y_i}}.
\eqn{X} is generated using distribution supplied by \code{rFUN_X}
(\code{rnorm} is the default).
Additional arguments to \code{rFUN_X} are supplied using the
\code{...} argument.
\eqn{M} and \eqn{Y} are generated using \eqn{X}
and the parameters provided using the regression equations above.
Residuals are assumed to be normaly distribution with means of zero
and provided variances
\eqn{\sigma^{2}_{\epsilon_{M}}} and
\eqn{\sigma^{2}_{\epsilon_{Y}}}.
}
\examples{
data <- gendat_med_simple(
  n = 100,
  alpha = 0.26^(1 / 2),
  tau_prime = 0,
  beta = 0.26^(1 / 2),
  delta_M = 0.490098,
  delta_Y = 0.490098,
  sigma2_epsilon_M = 0.74,
  sigma2_epsilon_Y = 0.74,
  rFUN_X = rnorm,
  mean = 0,
  sd = 1
)
}
\keyword{data}
