% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat.R
\name{gendat_mvn_fe}
\alias{gendat_mvn_fe}
\title{Generate Multivariate Normal Data from Variance-Covariance Matrix for Fixed-Effect Meta-Analysis.}
\usage{
gendat_mvn_fe(n, Sigma, mu = NULL, k, raw = TRUE, sumFUN = cor, ...)
}
\arguments{
\item{n}{Sample size.}

\item{Sigma}{\eqn{p \times p} variance-covariance matrix.}

\item{mu}{Mean vector. Defaults to zeros if unspecified.}

\item{k}{Number of primary studies.}

\item{raw}{Logical. Default is \code{TRUE}.
If \code{TRUE}, returns raw data.
If \code{FALSE}, returns summary function passed through \code{sumFUN}.}

\item{sumFUN}{Summary function used to summarize raw data.
By default data is summarized using the \code{cor} function.}

\item{...}{Arguments that can be passed to \code{MASS::mvrnorm}.}
}
\value{
Returns a list of length \code{k} of raw data or summary of the data
generated from a multivariate normal distribution
using the variance-covariance matrix and mean vector provided.
}
\description{
Generates multivariate normal data from
a \eqn{p \times p} variance-covariance matrix and
\eqn{p} dimensional vector of means.
\eqn{k} is used to define the number of primary studies
}
\examples{
n <- 100
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
mu <- c(100, 100, 100)
data <- gendat_mvn_fe(
  n = n,
  Sigma = Sigma,
  mu = mu,
  k = 20,
  raw = FALSE,
  sumFUN = cor
)
}
\seealso{
Other data generating functions: 
\code{\link{gendat_linreg_X}()},
\code{\link{gendat_linreg_y}()},
\code{\link{gendat_linreg}()},
\code{\link{gendat_mvn_a}()},
\code{\link{gendat_mvn}()},
\code{\link{gendat_vm}()},
\code{\link{gendat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{data generating functions}
\keyword{data}
