% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat.R
\name{gendat_percep}
\alias{gendat_percep}
\title{Generate Linearly Separable Data Based on the Perceptron Algorithm}
\usage{
gendat_percep(n, w, ...)
}
\arguments{
\item{n}{Sample size.}

\item{w}{A vector of weights defining the linear classifier.
The first element of the vector is the bias parameter \eqn{b}.}

\item{...}{Arguments to pass to \code{FUN}.}
}
\description{
Generates data that can be classified as \eqn{+1} or \eqn{-1}
based on a linear classifier defined by a vector of weights \eqn{w}.
}
\details{
Randomly generates the data matrix
\eqn{\mathbf{X}}
and vector
\eqn{\mathbf{y}}
using a threshold function:
if
\eqn{\mathbf{X}_{n \times k} \mathbf{w}_{k \times 1} > 0}
then \eqn{\mathbf{y}_{n \times 1} = +1} and
if
\eqn{\mathbf{X}_{n \times k} \mathbf{w}_{k \times 1} < 0}
then \eqn{\mathbf{y}_{n \times 1} = -1}.
}
\examples{
data <- gendat_percep(
  n = 100,
  w = c(.5, .5, .5),
  rFUN_X = rnorm,
  mean = 0,
  sd = 1
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{data}
