% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat.R
\name{gendat_vm}
\alias{gendat_vm}
\title{Generate Nonnormal Data Using the Vale and Maurelli (1983) Approach.}
\usage{
gendat_vm(
  n,
  Sigma_dot,
  skew,
  kurt,
  seed = sample(1:.Machine$integer.max, size = 1),
  rescale = FALSE,
  mu,
  sigma2
)
}
\arguments{
\item{n}{Sample size.}

\item{Sigma_dot}{\eqn{p \times p} correlation matrix (\eqn{\dot{\Sigma}}).}

\item{skew}{Skewness vector.
If a single value is given, the function assumes that
the \eqn{p} elements in the vector will have the same skewness.}

\item{kurt}{Kurtosis vector.
If a single value is given, the function assumes that
the \eqn{p} elements in the vector will have the same kurtosis.}

\item{seed}{Random seed for reproducibility.}

\item{rescale}{Logical.
Rescale the data using means \code{mu} (\eqn{\mu})
and variances \code{sigma2} (\eqn{\sigma^2}) provided.
Note that the actual means and variances that the final data set will have
are drawn randomly from a normal distrbution
using \code{mu} and \code{sigma2} as parameters.}

\item{mu}{Vector of means
corresponding to each element of \code{Sigma_dot} (\eqn{\dot{\Sigma}})
that will be used as population means when rescale is \code{TRUE}.
If a single value is given, the function assumes that
the \eqn{p} elements in the vector will have the same \code{mu}.}

\item{sigma2}{Vector of variances
corresponding to each element of \code{Sigma_dot} (\eqn{\dot{\Sigma}})
that will be used as population variances when rescale is \code{TRUE}.
If a single value is given, the function assumes that
the \eqn{p} elements in the vector will have the same \code{sigma2}.}
}
\value{
Returns an \eqn{n \times p} nonnormal data matrix
using the Vale and Maurelli (1983) approach
from a \eqn{p \times p} correlation matrix provided
and predefined skewness and kurtosis.
}
\description{
Generates nonnormal data using the Vale and Maurelli (1983)
approach from a correlation matrix \eqn{\dot{\Sigma}} with predefined skewness and kurtosis.
}
\examples{
Sigma_dot <- matrix(
  data = .60,
  ncol = 4,
  nrow = 4
)
diag(Sigma_dot) <- 1
n <- 100
skew <- 1.75
kurt <- 3.75
mu <- 100
sigma2 <- 225
data <- gendat_vm(
  n = n,
  Sigma_dot = Sigma_dot,
  skew = skew,
  kurt = kurt,
  mu = mu,
  sigma2 = sigma2
)
}
\references{
Fleishman, A. I (1978).
A method for simulating non-normal distributions.
\emph{Psychometrika, 43}, 521--532.

Vale, D. C., & Maurelli, V. A. (1983).
Simulating multivariate nonnormal distributions.
\emph{Psychometrika, 48}, 465--471.
}
\seealso{
Other data generating functions: 
\code{\link{gendat_linreg_X}()},
\code{\link{gendat_linreg_y}()},
\code{\link{gendat_linreg}()},
\code{\link{gendat_mvn_a}()},
\code{\link{gendat_mvn_fe}()},
\code{\link{gendat_mvn}()},
\code{\link{gendat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{data generating functions}
\keyword{data}
