% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heights}
\alias{heights}
\title{Mothers' and daughters' Heights}
\format{
A dataframe with 1375 rows and two variables
\describe{
\item{mheight}{Mother's height, in inches.}
\item{dheight}{Daughter's height, in inches.}
}
}
\usage{
data(heights)
}
\description{
Heights of mothers in the United Kingdom under the age of 65 and
one of their adult daughters over the age of 18
collected during the period 1893-1898.
}
\examples{
data(heights)
}
\references{
Pearson, E. S. and Lee, S. (1903). On the laws of inheritance in man. \emph{Biometrika}, 2, 357-–463.
}
\keyword{datasets}
