% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{inv_logit}
\alias{inv_logit}
\title{Sigmoid or Logistic Function}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
The inverse of the logit of \code{x} is given by
\eqn{\frac{1}{1 + \exp(-x)}}  or
\eqn{\frac{\exp\left( x \right)}{\exp \left( x \right) + 1}}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
