% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis}
\usage{
kurtosis(x, excess = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{excess}{Logical.
Return excess kurtosis
(kurtosis minus 3).}
}
\description{
The population Pearson's kurtosis is defined by
\deqn{\frac{\mu_4}{\left( \sigma^2 \right)^2} = \frac{\mu_4}{\sigma^4}}
where
\itemize{
\item \eqn{\mu_4} is fourth moment about the mean, and
\item \eqn{\sigma^2} is the second moment about the mean or the variance.
If we have sample data,
we substitute the population moments with the sample estimates.
\deqn{\frac{m_4}{\left( m_2 \right)^2}}
where
\item \eqn{m_4} is the fourth sample moment about the mean, and
\item \eqn{m_2} is the second sample moment about the mean or the sample variance.
Excess kurtosis (kurtosis minus 3) is an adjusted version of Pearson's kurtosis
to provide the comparison to the normal distribution.
Note that the normal distribution has an excess kurtosis of 3.
}
}
\references{
\href{https://en.wikipedia.org/wiki/Kurtosis}{Wikipedia: Kurtosis}
}
