% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.R
\name{lav}
\alias{lav}
\title{Generic SEM Fit Function}
\usage{
lav(
  data,
  model,
  raw_data = TRUE,
  mean_vector = NULL,
  n = NULL,
  minimal = TRUE,
  ...
)
}
\arguments{
\item{data}{If \code{raw_data = TRUE}, sample raw data.
The column names must contain observed variable names.
If \code{raw_data = FALSE}, numeric matrix.
A sample variance-covariance matrix.
The raw names and column names must contain observed variable names.}

\item{model}{Specified model to fit following \code{lavaan} notation.}

\item{raw_data}{Logical.
If \code{TRUE}, \code{data} is a sample raw data.
If \code{FALSE}, \code{data} is a sample variance-covariance matrix.}

\item{mean_vector}{A sample mean vector.
Ignored if \code{raw_data = TRUE}.}

\item{n}{Sample size.
Ignored if \code{raw_data = TRUE}.}

\item{minimal}{Logical.
If \code{TRUE}, returns the indirect effect of X on Y through M.
If \code{FALSE}, returns all the regression coefficients estimated.}

\item{...}{Arguments to pass to \code{lavaan::sem}.}
}
\description{
Fits a specified SEM model on given data using lavaan.
}
