% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression.}
\usage{
linreg(X, y, optimize = FALSE, ...)
}
\arguments{
\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}

\item{optimize}{Logical.
If \code{TRUE}, uses optimization to estimate parameters.
If \code{FALSE}, uses the closed form of the Ordinary Least Squares (OLS) estimator.}

\item{...}{Arguments to be passed to the optimization function specified.
This is only used when \code{optim} is \code{TRUE}}
}
\value{
If \code{optimize} is \code{TRUE}, returns the results of the optimization process.
If \code{optimize} is \code{FALSE}, returns a \eqn{k \times 1} matrix
of \eqn{k} unknown regression parameters estimated using ordinary least squares.
}
\description{
Estimates parameters \eqn{\boldsymbol{\hat{\beta}}} of a linear regression model
given by
\eqn{\mathbf{y_{n \time 1}} = \mathbf{X_{n \times k}b_{k \times i}} +
  \mathbf{e_{n \times 1}}}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
