% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_anova}
\alias{linreg_anova}
\title{Linear Regression (ANOVA)}
\usage{
linreg_anova(ess = NULL, rss = NULL, tss = NULL, n, k, ...)
}
\arguments{
\item{ess}{Explained sum of squares.
If \code{NULL},
\code{\link{linreg_ess}} is used to estimate \code{ess}.}

\item{rss}{Residual sum of squares.
If \code{NULL},
\code{\link{linreg_rss}} is used to estimate \code{rss}.}

\item{tss}{Total sum of squares.
If \code{NULL},
\code{\link{linreg_tss}} is used to estimate \code{tss}.}

\item{n}{Sample size.}

\item{k}{Number of regressors.}

\item{...}{Arguments to pass to
\code{\link{linreg_ess}},
\code{\link{linreg_rss}}, and
\code{\link{linreg_tss}}
if \code{ess}, \code{rss}, and \code{tss} are set to \code{NULL}.}
}
\value{
Returns elements of the ANOVA table.
}
\description{
Calculates the elements of the ANOVA table.
}
\author{
Ivan Jacob Agaloos Pesigan
}
