% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_e}
\alias{linreg_e}
\title{Linear Regression Residuals}
\usage{
linreg_e(beta_hat = NULL, X, y, m = FALSE)
}
\arguments{
\item{beta_hat}{Vector of \eqn{k} estimated regression parameters.
If \code{NULL},
regression coefficients are estimated using
\eqn{
    \left(
      \mathbf{X}^{\prime}
      \mathbf{X}
    \right)^{-1}
    \left(
      \mathbf{X}^{\prime}
      \mathbf{y}
    \right)
  }.}

\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}

\item{m}{Logical.
If \code{TRUE},
the function uses an alternative formula
\eqn{
    e
    =
    \mathbf{M}
    \mathbf{y}
  }.
See \code{\link{linreg_m}} for
\eqn{\mathbf{M}}.}
}
\value{
Returns the residuals.
}
\description{
Calculates the residuals
\deqn{
    \mathbf{e}
    =
    \mathbf{y}
    -
    \mathbf{X}
    \boldsymbol{\hat{\beta}}
  }
or
\deqn{
    e
    =
    \mathbf{M}
    \mathbf{y}
  }.
}
\author{
Ivan Jacob Agaloos Pesigan
}
