% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_ess}
\alias{linreg_ess}
\title{Linear Regression Explained Sum of Squares.}
\usage{
linreg_ess(beta_hat = NULL, X, y)
}
\arguments{
\item{beta_hat}{Vector of \eqn{k} estimated regression parameters.
If \code{NULL},
regression coefficients are estimated using
\eqn{
    \left(
      \mathbf{X}^{\prime}
      \mathbf{X}
    \right)^{-1}
    \left(
      \mathbf{X}^{\prime}
      \mathbf{y}
    \right)
  }.}

\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}
}
\value{
Returns the explained sum of squares.
}
\description{
Calculates explained sums of squares (ESS)
\deqn{
    \boldsymbol{\hat{\beta}}^{\prime}
    \mathbf{X}^{\prime}
    \mathbf{X}
    \boldsymbol{\hat{\beta}}
    -
    n
    \mathbf{\bar{Y}}^{2}
  }.
}
\author{
Ivan Jacob Agaloos Pesigan
}
