% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_h}
\alias{linreg_h}
\title{Linear Regression Hat Matrix}
\usage{
linreg_h(X)
}
\arguments{
\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}
}
\value{
Returns the hat matrix (\eqn{H}).
}
\description{
Calculates the hat matrix (H)
\deqn{
    \mathbf{H}
    =
    \mathbf{X}
    \left(
      \mathbf{X}^{\prime}
      \mathbf{X}
    \right)^{-1}
    \mathbf{X}^{\prime}}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
