% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_ols}
\alias{linreg_ols}
\title{Linear Regression (OLS).}
\usage{
linreg_ols(X, y, FUN = linreg_inv)
}
\arguments{
\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}

\item{FUN}{OLS function to use.
The options are \code{\link{linreg_inv}} using the inverse of \code{X},
\code{\link{linreg_qr}} using the QR Decomposition of \code{X}, and
\code{\link{linreg_svd}} using the Singular Value Decomposition (SVD) of \code{X}.}
}
\value{
Returns a \eqn{k \times 1} matrix of \eqn{k} unknown regression coefficients
(\eqn{\hat{\beta}})
estimated using ordinary least squares.
}
\description{
Estimates parameters of a linear regression model
using the closed form of the Ordinary Least Squares (OLS) estimator.
}
\author{
Ivan Jacob Agaloos Pesigan
}
