% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_qr}
\alias{linreg_qr}
\title{Linear Regression Coefficients (QR Decomposition).}
\usage{
linreg_qr(X, y)
}
\arguments{
\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}
}
\value{
Returns a \eqn{k \times 1} matrix of \eqn{k} unknown regression coefficients
(\eqn{\hat{\beta}})
estimated using ordinary least squares.
}
\description{
Estimates coefficients of a linear regression model
using the QR Decomposition.
}
