% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_r2}
\alias{linreg_r2}
\title{Linear Regression (R-square)}
\usage{
linreg_r2(beta_hat = NULL, X, y, m = FALSE, rss = TRUE)
}
\arguments{
\item{beta_hat}{Vector of \eqn{k} estimated regression parameters.
If \code{NULL},
regression coefficients are estimated using
\eqn{
    \left(
      \mathbf{X}^{\prime}
      \mathbf{X}
    \right)^{-1}
    \left(
      \mathbf{X}^{\prime}
      \mathbf{y}
    \right)
  }.}

\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}

\item{m}{Logical.
If \code{TRUE},
the function uses an alternative formula
\eqn{
    e
    =
    \mathbf{M}
    \mathbf{y}
  }.
See \code{\link{linreg_m}} for
\eqn{\mathbf{M}}.}

\item{rss}{Logical.
If \code{TRUE}, the function uses the residual sum of squares in the calculation.
If \code{FALSE}, the function uses the estimated sum of squares in the calculation.}
}
\value{
Returns the coefficient of determination \eqn{R^2}.
}
\description{
Calculates the coefficient of determination
(\eqn{R^2 = 1 - \frac{\textrm{Residual sum of squares}}{\textrm{Total sum of squares}}} or
\eqn{R^2 = \frac{\textrm{Explained sum of squares}}{\textrm{Total sum of squares}}}).
}
\author{
Ivan Jacob Agaloos Pesigan
}
