% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_rbar2}
\alias{linreg_rbar2}
\title{Linear Regression (Adjusted R-square)}
\usage{
linreg_rbar2(r2 = NULL, n, k, ...)
}
\arguments{
\item{r2}{Coefficient of determination \eqn{R^2}.}

\item{n}{Sample size.}

\item{k}{Number of regressors.}

\item{...}{Arguments to pass to \code{\link{linreg_r2}} if \code{r2 = NULL}.}
}
\value{
Returns the adjusted coefficient of determination \eqn{\bar{R}^{2}}.
}
\description{
Calculates the adjusted coefficient of determination
(\eqn{\bar{R}^{2} = 1 - \left( 1 - R^2 \right) \frac{n - 1}{n - k}}).
}
\author{
Ivan Jacob Agaloos Pesigan
}
