% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_rss}
\alias{linreg_rss}
\title{Linear Regression Residual Sum of Squares.}
\usage{
linreg_rss(beta_hat = NULL, X, y)
}
\arguments{
\item{beta_hat}{Vector of \eqn{k} estimated regression parameters.
If \code{NULL},
regression coefficients are estimated using
\eqn{
    \left(
      \mathbf{X}^{\prime}
      \mathbf{X}
    \right)^{-1}
    \left(
      \mathbf{X}^{\prime}
      \mathbf{y}
    \right)
  }.}

\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}
}
\value{
Returns the residual sum of squares.
}
\description{
Calculates residual sum of squares (RSS)
\deqn{
    \Sigma e_{i}^{2}
    =
    \Sigma_{i = 1}^{n}
    \left(
      y_i
      -
      \hat{y_i}
    \right)^2
    =
    \left(
      \mathbf{y}
      -
      \mathbf{X}
      \boldsymbol{\hat{\beta}}
    \right)^{\prime}
    \left(
      \mathbf{y}
      -
      \mathbf{X}
      \boldsymbol{\hat{\beta}}
    \right)
    =
    \mathbf{e^{\prime} e }
  },
where
\deqn{
    \mathbf{e}
    =
    \mathbf{y}
    -
    \mathbf{X}
    \boldsymbol{\hat{\beta}}
  }
or
\deqn{
    e
    =
    \mathbf{M}
    \mathbf{y}
  }.
}
\author{
Ivan Jacob Agaloos Pesigan
}
