% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_tss}
\alias{linreg_tss}
\title{Linear Regression Total Sum of Squares.}
\usage{
linreg_tss(y)
}
\arguments{
\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}
}
\value{
Returns the total sum of squares.
}
\description{
Calculates total sum of squares (TSS)
\deqn{
    \mathbf{y}^{\prime}
    \mathbf{y}
    -
    n
    \mathbf{\bar{Y}}^{2}
  }.
}
\author{
Ivan Jacob Agaloos Pesigan
}
