% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_yhat}
\alias{linreg_yhat}
\title{Linear Regression (y hat)}
\usage{
linreg_yhat(X, y)
}
\arguments{
\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}
}
\value{
Returns \eqn{\hat{y}}.
}
\description{
Calculates the y hat
(\eqn{\hat{y}
  = \mathbf{y} - \mathbf{e}
  = \mathbf{X} \boldsymbol{\hat{\beta}}
  = \mathbf{H} \mathbf{y} }).
}
\author{
Ivan Jacob Agaloos Pesigan
}
