% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisrel.R
\name{lisrel}
\alias{lisrel}
\title{LISREL Structural Equations with Latent Variables.}
\usage{
lisrel(LY, LX, TE, TD, BE, I, GA, PS, PH)
}
\arguments{
\item{LY}{\eqn{\boldsymbol{\Lambda}_{\mathbf{y}}}
\eqn{p \times m}
matrix of factor loadings
(\eqn{\boldsymbol{\lambda}}).
\eqn{p}
is the number of observed indicators
(\eqn{\mathbf{y}})
and
\eqn{m}
is the number of latent endogenous variables
(\eqn{\boldsymbol{\eta}}).}

\item{LX}{\eqn{\boldsymbol{\Lambda}_{\mathbf{x}}}
\eqn{q \times n}
matrix of factor loadings
(\eqn{\boldsymbol{\lambda}}).
\eqn{q}
is the number of observed indicators
(\eqn{\mathbf{x}})
and
\eqn{n}
is the number of latent exogenous variables
(\eqn{\boldsymbol{\xi}}).}

\item{TE}{\eqn{\boldsymbol{\Theta_{\boldsymbol{\epsilon}}}}
\eqn{p \times p}
matrix
of residual variances and covariances for
\eqn{\mathbf{y}}
(\eqn{\boldsymbol{\epsilon}}).}

\item{TD}{\eqn{\boldsymbol{\Theta_{\boldsymbol{\delta}}}}
\eqn{q \times q}
matrix
of residual variances and covariances for
\eqn{\mathbf{x}}
(\eqn{\boldsymbol{\delta}}).}

\item{BE}{\eqn{\mathbf{B}_{m \times m}}
coefficient matrix
for endogenous variables.}

\item{I}{\eqn{\mathbf{I}_{m \times m}}
identity matrix.}

\item{GA}{\eqn{\boldsymbol{\Gamma}_{m \times n}}
coefficient matrix
for exogenous variables.}

\item{PS}{\eqn{\boldsymbol{\Psi}_{m \times m}}
variance-covariance of
\eqn{\boldsymbol{\zeta}}.
\eqn{\boldsymbol{\zeta}}
is a
matrix of
residual variances and covariances in regression equations.}

\item{PH}{\eqn{\boldsymbol{\Phi}_{n \times n}}
variance-covariance matrix of
\eqn{\boldsymbol{\xi}}.}
}
\value{
Returns the model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
derived from the
\eqn{\boldsymbol{\Lambda}_{\mathbf{y}}}
(\code{LY}),
\eqn{\boldsymbol{\Lambda}_{\mathbf{x}}}
(\code{LX}),
\eqn{\boldsymbol{\Theta}_{\boldsymbol{\epsilon}}}
(\code{TE}),
\eqn{\boldsymbol{\Theta}_{\boldsymbol{\delta}}}
(\code{TD}),
\eqn{\mathbf{B}}
(\code{BE}),
\eqn{\mathbf{I}}
(\code{I}),
\eqn{\boldsymbol{\Gamma}}
(\code{GA}),
\eqn{\boldsymbol{\Psi}}
(\code{PS}),
and
\eqn{\boldsymbol{\Phi}}
(\code{PH})
matrices.
}
\description{
Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
using the LISREL notation
for structural equations with latent variables.
}
\details{
Combines
\eqn{\boldsymbol{\Sigma}_{\mathbf{yy}} \left( \boldsymbol{\theta} \right)}
(from \code{\link{lisrel_yy}}),
\eqn{\boldsymbol{\Sigma}_{\mathbf{yx}} \left( \boldsymbol{\theta} \right)}
(from \code{\link{lisrel_yx}}),
\eqn{\boldsymbol{\Sigma}_{\mathbf{xy}} \left( \boldsymbol{\theta} \right)}
(from \code{\link{lisrel_xy}}),
and
\eqn{\boldsymbol{\Sigma}_{\mathbf{xx}} \left( \boldsymbol{\theta} \right)}
(from \code{\link{lisrel_xx}})
to produce
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}
.
The variance-covariance matrices are derived using the following equations
\deqn{\boldsymbol{\Sigma}_{\mathbf{yy}} \left( \boldsymbol{\theta} \right)
  =
  \boldsymbol{\Lambda}_{\mathbf{y}}
  \left( \mathbf{I} - \mathbf{B} \right)^{-1}
  \left(
    \boldsymbol{\Gamma}
    \boldsymbol{\Phi}
    \boldsymbol{\Gamma}^{T} +
    \boldsymbol{\Psi}
  \right)
  \left[ \left( \mathbf{I} - \mathbf{B} \right)^{-1} \right]^{T}
  \boldsymbol{\Lambda}_{\mathbf{y}}^{T} +
  \boldsymbol{\Theta}_{\boldsymbol{\epsilon}}
  }
\deqn{\boldsymbol{\Sigma}_{\mathbf{yx}} \left( \boldsymbol{\theta} \right)
  =
  \boldsymbol{\Lambda}_{\mathbf{y}}
  \left( \mathbf{I} - \mathbf{B} \right)^{-1}
  \boldsymbol{\Gamma}
  \boldsymbol{\Phi}
  \boldsymbol{\Lambda}_{\mathbf{x}}^{T}
  }
\deqn{\boldsymbol{\Sigma}_{\mathbf{xy}} \left( \boldsymbol{\theta} \right)
  =
  \boldsymbol{\Lambda}_{\mathbf{x}}
  \boldsymbol{\Phi}
  \boldsymbol{\Gamma}^{T}
  \left[ \left( \mathbf{I} - \mathbf{B} \right)^{-1} \right]^{T}
  \boldsymbol{\Lambda}_{\mathbf{y}}^{T}
  }
\deqn{\boldsymbol{\Sigma}_{\mathbf{xx}} \left( \boldsymbol{\theta} \right)
  =
  \boldsymbol{\Lambda}_{\mathbf{x}}
  \boldsymbol{\Phi}
  \boldsymbol{\Lambda}_{\mathbf{y}}^{T} +
  \boldsymbol{\Theta}_{\boldsymbol{\delta}} .
  }
}
\references{
Bollen, K. A. (1989).
\emph{Structural equations with latent variables}.
New York: Wiley.

Jöreskog, K. G., & Sörbom, D. (1996).
\emph{Lisrel 8: User's reference guide} (2nd ed.).
Scientific Software.
}
\seealso{
Other SEM notation functions: 
\code{\link{eqs_mu}()},
\code{\link{eqs}()},
\code{\link{lisrel_fa}()},
\code{\link{lisrel_obs_xy}()},
\code{\link{lisrel_obs_yx}()},
\code{\link{lisrel_obs_yy}()},
\code{\link{lisrel_obs}()},
\code{\link{lisrel_xx}()},
\code{\link{lisrel_xy}()},
\code{\link{lisrel_yx}()},
\code{\link{lisrel_yy}()},
\code{\link{ram_mu}()},
\code{\link{ram_m}()},
\code{\link{ram_s}()},
\code{\link{ram}()},
\code{\link{sem_fa}()},
\code{\link{sem_lat}()},
\code{\link{sem_obs}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{lisrel}
\keyword{matrix}
