% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisrel.R
\name{lisrel_mat}
\alias{lisrel_mat}
\title{LISREL Matrices}
\usage{
lisrel_mat(LY, LX, TE, TD, eta_on_eta = NULL, eta_on_xi, PS, PH, latent = TRUE)
}
\arguments{
\item{LY}{\eqn{\boldsymbol{\Lambda}_{\mathbf{y}}}
\eqn{p \times m}
matrix of factor loadings
(\eqn{\boldsymbol{\lambda}}).
\eqn{p}
is the number of observed indicators
(\eqn{\mathbf{y}})
and
\eqn{m}
is the number of latent endogenous variables
(\eqn{\boldsymbol{\eta}}).}

\item{LX}{\eqn{\boldsymbol{\Lambda}_{\mathbf{x}}}
\eqn{q \times n}
matrix of factor loadings
(\eqn{\boldsymbol{\lambda}}).
\eqn{q}
is the number of observed indicators
(\eqn{\mathbf{x}})
and
\eqn{n}
is the number of latent exogenous variables
(\eqn{\boldsymbol{\xi}}).}

\item{TE}{\eqn{\boldsymbol{\Theta_{\boldsymbol{\epsilon}}}}
\eqn{p \times p}
matrix
of residual variances and covariances for
\eqn{\mathbf{y}}
(\eqn{\boldsymbol{\epsilon}}).
If a vector is supplied,
the matrix is assummed to be a diagonal matrix
with diagonals equal to \code{TE}.}

\item{TD}{\eqn{\boldsymbol{\Theta_{\boldsymbol{\delta}}}}
\eqn{q \times q}
matrix
of residual variances and covariances for
\eqn{\mathbf{x}}
(\eqn{\boldsymbol{\delta}}).
If a vector is supplied,
the matrix is assummed to be a diagonal matrix
with diagonals equal to \code{TD}.}

\item{eta_on_eta}{Vector values of the strict lower triangle of
\eqn{\mathbf{B}_{m \times m}}
(\code{BE}),
that is,
the coefficients
for endogenous variables.
Note that lower triangle is populated by column.
If no value for \code{eta_on_eta} is provided,
it is assumed that there are no direct paths
between any of the endogenous variables.}

\item{eta_on_xi}{Vector values of
\eqn{\boldsymbol{\Gamma}_{m \times n}}
coefficient matrix
for latent exogenous variables.
Note that matrix is populated by column.}

\item{PS}{\eqn{\boldsymbol{\Psi}_{m \times m}}
variance-covariance of
\eqn{\boldsymbol{\zeta}}.
\eqn{\boldsymbol{\zeta}}
is a
matrix of
residual variances and covariances in regression equations.
If a vector is supplied,
the matrix is assummed to be a diagonal matrix
with diagonals equal to \code{PS}.}

\item{PH}{\eqn{\boldsymbol{\Phi}_{n \times n}}
variance-covariance matrix of
\eqn{\boldsymbol{\xi}}.
If a vector is supplied,
the matrix is assummed to be a diagonal matrix
with diagonals equal to \code{PH}.}

\item{latent}{Logical.
If \code{TRUE}, uses the structural equations with latent variables LISREL notation,
If \code{FALSE}, uses the structural equations with observed variables LISREL notation.}
}
\description{
Helper function to construct LISREL Matrices
}
\author{
Ivan Jacob Agaloos Pesigan
}
