% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisrel.R
\name{lisrel_obs}
\alias{lisrel_obs}
\title{LISREL Structural Equations with Observed Variables.}
\usage{
lisrel_obs(BE, I, GA, PH, PS)
}
\arguments{
\item{BE}{\eqn{\mathbf{B}_{m \times m}}
coefficient matrix
for endogenous variables.}

\item{I}{\eqn{\mathbf{I}_{m \times m}}
identity matrix.}

\item{GA}{\eqn{\boldsymbol{\Gamma}_{m \times n}}
coefficient matrix
for exogenous variables.}

\item{PH}{\eqn{\boldsymbol{\Phi}_{n \times n}}
variance-covariance matrix of
\eqn{\boldsymbol{\xi}}.}

\item{PS}{\eqn{\boldsymbol{\Psi}_{m \times m}}
variance-covariance of
\eqn{\boldsymbol{\zeta}}.
\eqn{\boldsymbol{\zeta}}
is a
matrix of
residual variances and covariances in regression equations.}
}
\value{
Returns the model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
derived from the
\eqn{\mathbf{B}}
(\code{BE}),
\eqn{\mathbf{I}}
(\code{I}),
\eqn{\boldsymbol{\Gamma}}
(\code{GA}),
\eqn{\boldsymbol{\Phi}}
(\code{PH}),
and
\eqn{\boldsymbol{\Psi}}
(\code{PS})
matrices.
}
\description{
Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \theta \right)})
using the LISREL notation
for structural equations with observed variables.
}
\details{
Combines
\eqn{\boldsymbol{\Sigma}_{\mathbf{yy}} \left( \boldsymbol{\theta} \right)}
(\code{\link{lisrel_obs_yy}}),
\eqn{\boldsymbol{\Sigma}_{\mathbf{yx}} \left( \boldsymbol{\theta} \right)}
(\code{\link{lisrel_obs_yx}}),
\eqn{\boldsymbol{\Sigma}_{\mathbf{xy}} \left( \boldsymbol{\theta} \right)}
(\code{\link{lisrel_obs_xy}}), and
\eqn{\boldsymbol{\Sigma}_{\mathbf{xx}} \left( \boldsymbol{\theta} \right)}
(\eqn{\boldsymbol{\Phi}})
to produce
\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}.
}
\references{
Bollen, K. A. (1989).
\emph{Structural equations with latent variables}.
New York: Wiley.

Jöreskog, K. G., & Sörbom, D. (1996).
\emph{Lisrel 8: User's reference guide} (2nd ed.).
Scientific Software.
}
\seealso{
Other SEM notation functions: 
\code{\link{eqs_mu}()},
\code{\link{eqs}()},
\code{\link{lisrel_fa}()},
\code{\link{lisrel_obs_xy}()},
\code{\link{lisrel_obs_yx}()},
\code{\link{lisrel_obs_yy}()},
\code{\link{lisrel_xx}()},
\code{\link{lisrel_xy}()},
\code{\link{lisrel_yx}()},
\code{\link{lisrel_yy}()},
\code{\link{lisrel}()},
\code{\link{ram_mu}()},
\code{\link{ram_m}()},
\code{\link{ram_s}()},
\code{\link{ram}()},
\code{\link{sem_fa}()},
\code{\link{sem_lat}()},
\code{\link{sem_obs}()}
}
\concept{SEM notation functions}
\keyword{lisrel}
\keyword{matrix}
