% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisrel.R
\name{lisrel_obs_xy}
\alias{lisrel_obs_xy}
\title{LISREL Structural Equations with Observed Variables (xy)}
\usage{
lisrel_obs_xy(inv, GA, PH)
}
\arguments{
\item{inv}{The inverse of \code{I} minus \code{BE}
(\eqn{\left( \mathbf{I} - \mathbf{B} \right)^{-1}}).}

\item{GA}{\eqn{\boldsymbol{\Gamma}_{m \times n}}
coefficient matrix
for exogenous variables.}

\item{PH}{\eqn{\boldsymbol{\Phi}_{n \times n}}
variance-covariance matrix of
\eqn{\boldsymbol{\xi}}.}
}
\value{
Returns the model-implied variance-covariance matrix for
\eqn{\mathbf{xy}}
(\eqn{\boldsymbol{\Sigma}_{\mathbf{xy}} \left( \boldsymbol{\theta} \right)})
derived from the
\eqn{\mathbf{B}}
(\code{BE}),
\eqn{\mathbf{I}}
(\code{I}),
\eqn{\boldsymbol{\Gamma}}
(\code{GA}),
and
\eqn{\boldsymbol{\Phi}}
(\code{PH}),
matrices.
}
\description{
Model-implied variance-covariance matrix for \eqn{\mathbf{y}} variables
(\eqn{\boldsymbol{\Sigma}_{\mathbf{xy}} \left( \boldsymbol{\theta} \right)})
using the LISREL notation
for structural equations with observed variables.
}
\details{
\deqn{\boldsymbol{\Sigma}_{\mathbf{xy}} \left( \boldsymbol{\theta} \right)
  =
  \left( \mathbf{I} - \mathbf{B} \right)^{-1}
  \left(
    \boldsymbol{\Gamma}
    \boldsymbol{\Phi}
    \boldsymbol{\Gamma}^{T} +
    \boldsymbol{\Psi}
  \right)
  \left[ \left( \mathbf{I} - \mathbf{B} \right)^{-1} \right]^{T}
  }
}
\references{
Bollen, K. A. (1989).
\emph{Structural equations with latent variables}.
New York: Wiley.

Jöreskog, K. G., & Sörbom, D. (1996).
\emph{Lisrel 8: User's reference guide} (2nd ed.).
Scientific Software.
}
\seealso{
Other SEM notation functions: 
\code{\link{eqs_mu}()},
\code{\link{eqs}()},
\code{\link{lisrel_fa}()},
\code{\link{lisrel_obs_yx}()},
\code{\link{lisrel_obs_yy}()},
\code{\link{lisrel_obs}()},
\code{\link{lisrel_xx}()},
\code{\link{lisrel_xy}()},
\code{\link{lisrel_yx}()},
\code{\link{lisrel_yy}()},
\code{\link{lisrel}()},
\code{\link{ram_mu}()},
\code{\link{ram_m}()},
\code{\link{ram_s}()},
\code{\link{ram}()},
\code{\link{sem_fa}()},
\code{\link{sem_lat}()},
\code{\link{sem_obs}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{lisrel}
\keyword{matrix}
