% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisrel.R
\name{lisrel_residual}
\alias{lisrel_residual}
\title{Lisrel Residuals}
\usage{
lisrel_residual(obs, latent = TRUE, ...)
}
\arguments{
\item{obs}{Observed sample variance-covariance matrix.}

\item{latent}{Logical.
If \code{TRUE}, uses the structural equations with latent variables LISREL notation,
If \code{FALSE}, uses the structural equations with observed variables LISREL notation.}

\item{...}{Arguments to pass to \code{\link{lisrel_mat}}.}
}
\description{
Lisrel Residuals
}
