% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisrel.R
\name{lisrel_xx}
\alias{lisrel_xx}
\title{LISREL Structural Equations with Latent Variables (xx).}
\usage{
lisrel_xx(LX, TD, PH)
}
\arguments{
\item{LX}{\eqn{\boldsymbol{\Lambda}_{\mathbf{x}}}
\eqn{q \times n}
matrix of factor loadings
(\eqn{\boldsymbol{\lambda}}).
\eqn{q}
is the number of observed indicators
(\eqn{\mathbf{x}})
and
\eqn{n}
is the number of latent exogenous variables
(\eqn{\boldsymbol{\xi}}).}

\item{TD}{\eqn{\boldsymbol{\Theta_{\boldsymbol{\delta}}}}
\eqn{q \times q}
matrix
of residual variances and covariances for
\eqn{\mathbf{x}}
(\eqn{\boldsymbol{\delta}}).}

\item{PH}{\eqn{\boldsymbol{\Phi}_{n \times n}}
variance-covariance matrix of
\eqn{\boldsymbol{\xi}}.}
}
\value{
Returns the model-implied variance-covariance matrix for
\eqn{\mathbf{x}}
(\eqn{\boldsymbol{\Sigma}_{\mathbf{xx}} \left( \boldsymbol{\theta} \right)})
derived from the
\eqn{\boldsymbol{\Lambda}_{\mathbf{x}}}
(\code{LX}),
\eqn{\boldsymbol{\Theta}_{\boldsymbol{\delta}}}
(\code{TD}),
and
\eqn{\boldsymbol{\Phi}}
(\code{PH})
matrices.
}
\description{
Model-implied variance-covariance matrix for \eqn{\mathbf{x}} variables
(\eqn{\boldsymbol{\Sigma}_{\mathbf{xx}} \left( \boldsymbol{\theta} \right)})
using the LISREL notation
for structural equations with latent variables.
}
\details{
\deqn{\boldsymbol{\Sigma}_{\mathbf{xx}} \left( \boldsymbol{\theta} \right)
  =
  \boldsymbol{\Lambda}_{\mathbf{x}}
  \boldsymbol{\Phi}
  \boldsymbol{\Lambda}_{\mathbf{y}}^{T} +
  \boldsymbol{\Theta}_{\boldsymbol{\delta}}
}
}
\references{
Bollen, K. A. (1989).
\emph{Structural equations with latent variables}.
New York: Wiley.

Jöreskog, K. G., & Sörbom, D. (1996).
\emph{Lisrel 8: User's reference guide} (2nd ed.).
Scientific Software.
}
\seealso{
Other SEM notation functions: 
\code{\link{eqs_mu}()},
\code{\link{eqs}()},
\code{\link{lisrel_fa}()},
\code{\link{lisrel_obs_xy}()},
\code{\link{lisrel_obs_yx}()},
\code{\link{lisrel_obs_yy}()},
\code{\link{lisrel_obs}()},
\code{\link{lisrel_xy}()},
\code{\link{lisrel_yx}()},
\code{\link{lisrel_yy}()},
\code{\link{lisrel}()},
\code{\link{ram_mu}()},
\code{\link{ram_m}()},
\code{\link{ram_s}()},
\code{\link{ram}()},
\code{\link{sem_fa}()},
\code{\link{sem_lat}()},
\code{\link{sem_obs}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{lisrel}
\keyword{matrix}
