% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{logreg}
\alias{logreg}
\title{Logistic Regression}
\usage{
logreg(X, y, ...)
}
\arguments{
\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}

\item{...}{Arguments to be passed to the optimization function specified.
This is only used when \code{optim} is \code{TRUE}}
}
\description{
Estimates parameters \eqn{\mathbf{b}} of a logistic regression model
given by
\eqn{\mathbf{y_{n \time 1}} = \mathbf{X_{n \times k}b_{k \times 1}} +
  \mathbf{e_{n \times 1}}}
where \eqn{\mathbf{X_{n \times k}b_{k \times 1}}
  = \ln \left( \frac{\mu}{1 - \mu} \right)} and
\eqn{\mu = \frac{1}{1 + \exp(-1)} =
  \frac{\exp\left( x \right)}{\exp \left( x \right) + 1}}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
