% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{logreg_negll}
\alias{logreg_negll}
\title{Logistic Regression (-LL).}
\usage{
logreg_negll(theta, X, y)
}
\arguments{
\item{theta}{vector of estimated logistic regression parameters.}

\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}
}
\description{
Calculates the negative log-likelihood of \eqn{y}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
