% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{logreg_optim}
\alias{logreg_optim}
\title{Logistic regression (Optimization).}
\usage{
logreg_optim(X, y, start_values, optim = TRUE, ...)
}
\arguments{
\item{X}{The data matrix,
that is an \eqn{n \times k} matrix of \eqn{n} observations
of \eqn{k} regressors,
which includes a regressor whose value is 1 for each observation.}

\item{y}{\eqn{n \times 1} vector of observations on the regressand variable.}

\item{start_values}{A vector of starting values
corresponding to the number of parameters to be estimated.}

\item{optim}{Logical.
If \code{TRUE}, the \code{optim} function is used.
If \code{FALSE}, the \code{nlminb} function is used.}

\item{...}{Arguments to be passed to the optimization function specified.
This is only used when \code{optim} is \code{TRUE}}
}
\description{
Estimates parameters of a logistic regression model
using optimization.
}
\author{
Ivan Jacob Agaloos Pesigan
}
