% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{m4}
\alias{m4}
\title{Fourth Central Moment}
\usage{
m4(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
The sample fourth central moment is defined by
\deqn{m_4 = \frac{\sum_{i = 1}^{n} \left( x_i - \bar{x} \right)^4}{n}}
where
\itemize{
\item \eqn{n} is the sample size,
\item \eqn{x = \{x_1 \dots x_n\}} is a random variable, and
\item \eqn{\bar{x}} is the sample mean of \eqn{x}.
}
}
\references{
\href{https://en.wikipedia.org/wiki/Central_moment}{Wikipedia: Central Moment}
}
\author{
Ivan Jacob Agaloos Pesigan
}
