% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.R
\name{med_serial2}
\alias{med_serial2}
\title{Serial Mediation Model with Two Mediators}
\usage{
med_serial2(data, scale = FALSE, minimal = TRUE, s2 = FALSE, s2_est = "both")
}
\arguments{
\item{data}{A matrix with variables X, M1, M2, and Y.}

\item{scale}{Logical.
If \code{TRUE}, scales the data before fitting the model.}

\item{minimal}{Logical.
If \code{TRUE}, returns the indirect effect of X on Y through M1 and M2.
If \code{FALSE}, returns all the regression coefficients estimated.}

\item{s2}{Logical.
If \code{TRUE}, estimates residual variance.}

\item{s2_est}{String.
Residual variance estimator.
If \code{"both"},
returns both OLS and ML estimates as a vector.
If \code{"ols"},
returns OLS estimate.
If \code{"ml"},
returns ML estimate.
Ignored if \code{s2 = FALSE}.}
}
\description{
Estimates the indirect effect in a serial mediation model,
that is the product of \eqn{\alpha1}, \eqn{xi} and \eqn{\beta2} from
\eqn{M1_i = \delta_{M1} + \alpha1 X_i + \epsilon_{M1_i}},
\eqn{M2_i = \delta_{M2} + \alpha2 X_i + \xi M1_i + \epsilon_{M2_i}}, and
\eqn{Y_i = \delta_Y + \tau^{\prime} X_i + \beta1 M1_i + \beta2 M2_i + \epsilon_{Y_i}}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
