% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.R
\name{med_serial3_lav}
\alias{med_serial3_lav}
\title{Serial Mediation Model with Three Mediators (lavaan)}
\usage{
med_serial3_lav(data, scale = FALSE, minimal = TRUE, est = FALSE, ...)
}
\arguments{
\item{data}{A matrix with variables X, M, and Y.}

\item{scale}{Logical.
If \code{TRUE}, scales the data before fitting the model.}

\item{minimal}{Logical.
If \code{TRUE}, returns the indirect effect of X on Y through M.
If \code{FALSE}, returns all the regression coefficients estimated.}

\item{est}{Logical.
If \code{TRUE}, returns a vector of parameter estimates and standard errors.
If \code{FALSE}, returns the \code{lavaan} object.}

\item{...}{Arguments to pass to \code{lavaan::sem}.}
}
\description{
Estimates the indirect effect in a serial mediation model,
that is the product of \eqn{\alpha1}, \eqn{xi1} , \eqn{xi3}, and \eqn{\beta2} from
\eqn{M1_i = \delta_{M1} + \alpha1 X_i + \epsilon_{M1_i}},
\eqn{M2_i = \delta_{M2} + \alpha2 X_i + \xi1 M1_i + \epsilon_{M2_i}},
\eqn{M3_i = \delta_{M3} + \alpha3 X_i + \xi2 M1_i + \xi3 M2_i + \epsilon_{M3_i}}, and
\eqn{Y_i = \delta_Y + \tau^{\prime} X_i + \beta1 M1_i + \beta2 M2_i + \beta3 M3_i + \epsilon_{Y_i}}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
