% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.R
\name{med_serial3_ml}
\alias{med_serial3_ml}
\title{Serial Mediation Model with Three Mediators (optim)}
\usage{
med_serial3_ml(obs)
}
\arguments{
\item{obs}{Observed sample variance-covariance matrix.}
}
\description{
Serial Mediation Model with Three Mediators (optim)
}
\author{
Ivan Jacob Agaloos Pesigan
}
