% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.R
\name{med_simple}
\alias{med_simple}
\title{Simple Mediation Model}
\usage{
med_simple(data, scale = FALSE, minimal = TRUE, s2 = FALSE, s2_est = "both")
}
\arguments{
\item{data}{A matrix with variables X, M, and Y.}

\item{scale}{Logical.
If \code{TRUE}, scales the data before fitting the model.}

\item{minimal}{Logical.
If \code{TRUE}, returns the indirect effect of X on Y through M.
If \code{FALSE}, returns all the regression coefficients estimated.}

\item{s2}{Logical.
If \code{TRUE}, estimates residual variance.}

\item{s2_est}{String.
Residual variance estimator.
If \code{"both"},
returns both OLS and ML estimates as a vector.
If \code{"ols"},
returns OLS estimate.
If \code{"ml"},
returns ML estimate.
Ignored if \code{s2 = FALSE}.}
}
\description{
Estimates the indirect effect in a simple mediation model,
that is the product of \eqn{\alpha} and \eqn{\beta} from
\eqn{M_i = \delta_M + \alpha X_i + \epsilon_{M_i}} and
\eqn{Y_i = \delta_Y + \tau^{\prime} X_i + \beta M_i + \epsilon_{Y_i}}.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
mu <- c(100, 100, 100)
data <- gendat_mvn(
  n = 100,
  Sigma = Sigma,
  mu = mu
)
med_simple(data = data, minimal = FALSE, s2 = TRUE)
}
\author{
Ivan Jacob Agaloos Pesigan
}
