% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.R
\name{med_simple_ml}
\alias{med_simple_ml}
\title{Simple Mediation Model (optim)}
\usage{
med_simple_ml(obs)
}
\arguments{
\item{obs}{Observed sample variance-covariance matrix.}
}
\description{
Estimates the indirect effect in a simple mediation model,
that is the product of \eqn{\alpha} and \eqn{\beta} from
\eqn{M_i = \delta_M + \alpha X_i + \epsilon_{M_i}} and
\eqn{Y_i = \delta_Y + \tau^{\prime} X_i + \beta M_i + \epsilon_{Y_i}}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
