% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moment}
\alias{moment}
\title{Central Moment}
\usage{
moment(x, j)
}
\arguments{
\item{x}{Numeric vector.}

\item{j}{Integer.
\code{j}th moment.
From 0 to 4.}
}
\description{
The sample central moment is defined by
\deqn{m_j = \frac{\sum_{i = 1}^{n} \left( x_i - \bar{x} \right)^j}{n}}
where
\itemize{
\item \eqn{n} is the sample size,
\item \eqn{x = \{x_1 \dots x_n\}} is a random variable,
\item \eqn{\bar{x}} is the sample mean of \eqn{x}, and
\item \eqn{j} is the \eqn{j}th central moment.
}
}
\details{
\itemize{
\item The "zeroth" central moment is 1.
\item The first central moment is 0.
\item The second cental moment is the variance.
\item The third central moment is used to define skewness.
\item The fourth central moment is used to define kurtosis.
}
}
\references{
\href{https://en.wikipedia.org/wiki/Central_moment}{Wikipedia: Central Moment}
}
\author{
Ivan Jacob Agaloos Pesigan
}
