% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{opt}
\alias{opt}
\title{Estimation via Optimization}
\usage{
opt(FUN, start_values, optim = TRUE, ...)
}
\arguments{
\item{FUN}{Objective function.}

\item{start_values}{A vector of starting values
corresponding to the number of parameters to be estimated.}

\item{optim}{Logical.
If \code{TRUE}, the \code{optim} function is used.
If \code{FALSE}, the \code{nlminb} function is used.}

\item{...}{Additional arguments to be passed to the optimizer
and arguments to be passed to the objective function.}
}
\value{
Returns the output of \code{\link{optim}}
or \code{\link{nlminb}}.
}
\description{
Estimates parameters from a given objective function
using optimization.
}
\author{
Ivan Jacob Agaloos Pesigan
}
