% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{quad}
\alias{quad}
\title{Solve Quadratic Equation}
\usage{
quad(a, b, c)
}
\arguments{
\item{a}{Numeric value for quadratic term of x.}

\item{b}{Numeric value for multiplicative term of x.}

\item{c}{Numeric value for constant term.}
}
\value{
Returns the root/s fo the quadratic equation.
}
\description{
General function to solve classic quadratic equation:
\deqn{a x^2 + b x + c = 0}
from https://rpubs.com/kikihatzistavrou/80124
}
