% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram}
\alias{ram}
\title{RAM}
\usage{
ram(A, S, F, I)
}
\arguments{
\item{A}{Asymmetric paths,
such as regression coefficients and factor loadings.}

\item{S}{Symmetric matrix
representing variances and covariances.}

\item{F}{Filter matrix
used to select the observed variables.}

\item{I}{Identity matrix.}
}
\value{
Returns the model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
derived from the \eqn{\mathbf{A}},
\eqn{\mathbf{S}},
\eqn{\mathbf{F}}, and
\eqn{\mathbf{I}} matrices.
}
\description{
Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
using the Reticular Action Model notation.
}
\details{
\deqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)
  =
  \mathbf{F}
  \left( \mathbf{I} - \mathbf{A} \right)^{-1}
  \mathbf{S}
  \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{T}
  \mathbf{F}^{T}
}
}
\examples{
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
S <- F <- I <- diag(3)
S[1, 1] <- 225
S[2, 2] <- 166.5
S[3, 3] <- 166.5
ram(A = A, S = S, F = F, I = I)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology, 37}(2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ram_mu}()},
\code{\link{ram_m}()},
\code{\link{ram_s}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
