% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_residuals}
\alias{ram_residuals}
\title{RAM Residuals}
\usage{
ram_residuals(obs, A, S, F, I)
}
\arguments{
\item{obs}{Observed sample variance-covariance matrix.}

\item{A}{Asymmetric paths,
such as regression coefficients and factor loadings.}

\item{S}{Symmetric matrix
representing variances and covariances.}

\item{F}{Filter matrix
used to select the observed variables.}

\item{I}{Identity matrix.}
}
\description{
RAM Residuals
}
