% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{relu}
\alias{relu}
\title{Rectified Linear Unit Function}
\usage{
relu(x)
}
\arguments{
\item{x}{Input value.}
}
\description{
The rectified linear unit of \code{x} is given by
x < 0 returns 0
x >= 0 returns x
}
\author{
Ivan Jacob Agaloos Pesigan
}
