% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_fit.R
\name{sem_e}
\alias{sem_e}
\title{SEM (Residuals)}
\usage{
sem_e(obs, FUN = ram_residuals, ...)
}
\arguments{
\item{obs}{Observed sample variance-covariance matrix.}

\item{FUN}{Function to derive model implied variance-covariance matrix. The default is \code{ram_residuals}.}

\item{...}{Arguments to pass to \code{FUN}.}
}
\value{
Returns the residuals, that is, the difference between the observed and implied variance-covariance matrix.
}
\description{
SEM (Residuals)
}
