% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_obj_fn.R
\name{sem_fml}
\alias{sem_fml}
\title{Maximum Likelihood Objective Function}
\usage{
sem_fml(imp, obs)
}
\arguments{
\item{imp}{Model-implied variance-covariance matrix.}

\item{obs}{Observed sample variance-covariance matrix.}
}
\description{
Calculates the maximum likelihood objective function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
