% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_obj_fn.R
\name{sem_fuls}
\alias{sem_fuls}
\title{Unweighted Least Squares Objective Function}
\usage{
sem_fuls(imp, obs)
}
\arguments{
\item{imp}{Model-implied variance-covariance matrix.}

\item{obs}{Observed sample variance-covariance matrix.}
}
\description{
Calculates the unweighted least squares objective function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
