% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_other.R
\name{sem_lat}
\alias{sem_lat}
\title{Structural Equations with Latent Variables}
\usage{
sem_lat(L, TH, BE, I, PS)
}
\arguments{
\item{L}{\eqn{\boldsymbol{\Lambda}_{p \times m}}
matrix of factor loadings
(\eqn{\boldsymbol{\lambda}}).
\eqn{p}
is the number of indicators and
\eqn{m}
is the number of latent factor variables.}

\item{TH}{\eqn{\boldsymbol{\Theta}_{p \times p}}
matrix of residual variances and covariances.}

\item{BE}{\eqn{\mathbf{B}_{m \times m}}
coefficient matrix.}

\item{I}{\eqn{\mathbf{I}_{m \times m}}
identity matrix.}

\item{PS}{\eqn{\boldsymbol{\Psi}_{m \times m}}
variance-covariance matrix of
exogenous variables
as well as
variance-covariance of
endogenous variables.}
}
\description{
Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
for structural equations with latent variables.
}
\details{
\deqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)
  =
  \boldsymbol{\Lambda}
  \left( \mathrm{I} - \mathrm{B} \right)^{-1}
  \boldsymbol{\Psi}
  \left[ \left( \mathrm{I} - \mathrm{B} \right)^{-1} \right]^{T}
  \boldsymbol{\Lambda}^{T} +
  \boldsymbol{\Theta}
  }
Note that this notation,
known as the "All-y" notation,
treats
all variables as endogenous variables.
As such,
all latent variables are contained in
\eqn{\boldsymbol{\eta}}
and
all observed variables are contained in
\eqn{\mathbf{y}}.
It is important to take note that
\eqn{\boldsymbol{\Psi}}
contains both exogenous variances and covariances
as well as latent residual variances and covariances.
Note that
\eqn{\boldsymbol{\Psi}}
matrix combines
\eqn{\boldsymbol{\Psi}}
and
\eqn{\boldsymbol{\Phi}}
from the original LISREL notation
for structural equations with latent variables.
As such,
it contains
the variance-covariance of exogenous variables,
and residual variance-covariance of endogenous variables.
Note that
\eqn{\mathbf{B}}
matrix combines
\eqn{\mathbf{B}}
and
\eqn{\boldsymbol{\Gamma}}
from the original LISREL notation
for structural equations with latent variables.
As such,
it contains all coefficients.
}
\seealso{
Other SEM notation functions: 
\code{\link{eqs_mu}()},
\code{\link{eqs}()},
\code{\link{lisrel_fa}()},
\code{\link{lisrel_obs_xy}()},
\code{\link{lisrel_obs_yx}()},
\code{\link{lisrel_obs_yy}()},
\code{\link{lisrel_obs}()},
\code{\link{lisrel_xx}()},
\code{\link{lisrel_xy}()},
\code{\link{lisrel_yx}()},
\code{\link{lisrel_yy}()},
\code{\link{lisrel}()},
\code{\link{ram_mu}()},
\code{\link{ram_m}()},
\code{\link{ram_s}()},
\code{\link{ram}()},
\code{\link{sem_fa}()},
\code{\link{sem_obs}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
