% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_fit.R
\name{sem_log_h0}
\alias{sem_log_h0}
\title{SEM (Loglikelihood H0)}
\usage{
sem_log_h0(obs, imp, n)
}
\arguments{
\item{obs}{Observed sample variance-covariance matrix.}

\item{imp}{Model-implied variance-covariance matrix.}

\item{n}{Sample size.}
}
\description{
SEM (Loglikelihood H0)
}
