% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_fit.R
\name{sem_log_h1}
\alias{sem_log_h1}
\title{SEM (Loglikelihood H1)}
\usage{
sem_log_h1(obs, n)
}
\arguments{
\item{obs}{Observed sample variance-covariance matrix.}

\item{n}{Sample size.}
}
\description{
SEM (Loglikelihood H1)
}
