% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_fit.R
\name{sem_neg2ll}
\alias{sem_neg2ll}
\title{SEM (-2 Loglikelihood)}
\usage{
sem_neg2ll(obs = NULL, imp = NULL, n = NULL, log_h0 = NULL, log_h1 = NULL)
}
\arguments{
\item{obs}{Observed sample variance-covariance matrix.}

\item{imp}{Model-implied variance-covariance matrix.}

\item{n}{Sample size.}

\item{log_h0}{Loglikelihood H0.}

\item{log_h1}{Loglikelihood H0.}
}
\description{
Calculates -2 loglikelihood.
If
\code{log_h0}
and
\code{log_h1}
are \emph{provided},
\code{obs},
\code{imp},
and
\code{n}
are \emph{not needed}.
If
\code{log_h0}
and
\code{log_h1}
are \emph{not provided},
\code{obs},
\code{imp},
and
\code{n}
are \emph{needed}.
}
