% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_other.R
\name{sem_obs}
\alias{sem_obs}
\title{Structural Equations with Observed Variables}
\usage{
sem_obs(BE, I, PS)
}
\arguments{
\item{BE}{\eqn{\mathbf{B}_{m \times m}}
coefficient matrix.}

\item{I}{\eqn{\mathbf{I}_{m \times m}}
identity matrix.}

\item{PS}{\eqn{\boldsymbol{\Psi}_{m \times m}}
variance-covariance matrix
(variance-covariance of exogenous variables,
and residual variance-covariance of endogenous variables).}
}
\value{
Returns the model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
derived from the
\eqn{\mathbf{B}},
\eqn{\mathbf{I}},
and
\eqn{\boldsymbol{\Psi}}
matrices.
}
\description{
Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \theta \right)})
for structural equations with observed variables.
}
\details{
\eqn{\boldsymbol{\Sigma} \left( \theta \right)
  =
  \left( \mathbf{I} - \mathbf{B} \right)^{-1}
  \boldsymbol{\Psi}
  \left[ \left( \mathbf{I} - \mathbf{B} \right)^{-1} \right]^{T}
}
}
\examples{
BE <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
PS <- F <- I <- diag(3)
PS[1, 1] <- 225
PS[2, 2] <- 166.5
PS[3, 3] <- 166.5
sem_obs(BE = BE, I = I, PS = PS)
}
\references{
Bollen, K. A. (1989).
\emph{Structural equations with latent variables}.
New York: Wiley.

Jöreskog, K. G., & Sörbom, D. (1996).
\emph{Lisrel 8: User's reference guide} (2nd ed.).
Scientific Software.
}
\seealso{
Other SEM notation functions: 
\code{\link{eqs_mu}()},
\code{\link{eqs}()},
\code{\link{lisrel_fa}()},
\code{\link{lisrel_obs_xy}()},
\code{\link{lisrel_obs_yx}()},
\code{\link{lisrel_obs_yy}()},
\code{\link{lisrel_obs}()},
\code{\link{lisrel_xx}()},
\code{\link{lisrel_xy}()},
\code{\link{lisrel_yx}()},
\code{\link{lisrel_yy}()},
\code{\link{lisrel}()},
\code{\link{ram_mu}()},
\code{\link{ram_m}()},
\code{\link{ram_s}()},
\code{\link{ram}()},
\code{\link{sem_fa}()},
\code{\link{sem_lat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
