% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_fit.R
\name{sem_r2}
\alias{sem_r2}
\title{SEM (R square)}
\usage{
sem_r2(PS, Sigma_yy)
}
\arguments{
\item{PS}{\eqn{\boldsymbol{\Psi}_{m \times m}}
variance-covariance of
\eqn{\boldsymbol{\zeta}}.
\eqn{\boldsymbol{\zeta}}
is a
matrix of
residual variances and covariances in regression equations.}

\item{Sigma_yy}{Variance-covariance matrix of y variables
(\eqn{\Sigma_{yy}}).}
}
\description{
SEM (R square)
}
