% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wages}
\alias{wages}
\title{Wages of Workers}
\format{
A dataframe with 1289 rows and seven variables
\describe{
\item{wages}{Hourly wages in dollars.}
\item{gender}{1 for female, 0 for male.}
\item{race}{1 for nonwhite workers and 0 for white workers.}
\item{union}{1 if in union, 0 otherwise.}
\item{edu}{Education in years.}
\item{exp}{Work experience in years.}
\item{age}{Age in years.}
}
}
\usage{
data(wages)
}
\description{
Wages and related data for 1289 workers.
}
\examples{
data(wages)
}
\keyword{datasets}
