% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wages2}
\alias{wages2}
\title{Wages of Youth}
\format{
A dataframe with 540 rows and 16 variables
\describe{
\item{age}{Age in 2002.}
\item{school}{Years of schooling (highest grade completed as of 2002).}
\item{female}{1 for female, 0 for male.}
\item{male}{1 for female, 0 for male.}
\item{black}{1 for blacks.}
\item{hisp}{1 for Hispanic; non-black, non-Hispanic is the reference category.}
\item{exp}{Total out-of-school work experience in years as of the 2002 interview.}
\item{married}{1 if married, 0 otherwise.}
\item{schoolf}{Years of schooling of respondent's father.}
\item{schoolm}{Years of schooling of respondent's mother.}
\item{siblings}{Number of siblings.}
\item{wages}{Hourly wages in dollars.}
\item{asvab02}{Arithmetic reasoning.}
\item{asvab03}{Word knowledge.}
\item{lnwages}{Natural logarithm of wages.}
\item{exp2}{Work experience squared.}
}
}
\usage{
data(wages2)
}
\description{
Wages and educational attainment of youth in the USA
}
\examples{
data(wages2)
}
\keyword{datasets}
