% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{water}
\alias{water}
\title{Effects of Temperature on Water Consumption}
\format{
A dataframe with 50 rows and three variables
\describe{
\item{temp}{Room temperature in degrees Fahrenheit.
The temperature was manipulated such that the participants were exposed
to a specific temperature in the room for the 4 hours of the experiment.}
\item{thirst}{Self-reported measure of thirst at the end of a 2-hour period. The scale is from 1 (not at all thirsty) to 5 (very thirsty)}
\item{water}{The number of deciliters of water consumed during the last 2 hours of the study.}
}
}
\usage{
data(water)
}
\description{
An experimental study of the  effects of temperature on water consumption
through self-reported thirst.
Fifty participants were in a room for 4 hours doing a variety of tasks.
Before the experiment,
each participant was acclimated to a standard temperature
of 70 degrees Fahrenheit.
}
\examples{
data(water)
}
\keyword{datasets}
