% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_misc.R
\name{zero_hit}
\alias{zero_hit}
\title{Zero Hit}
\usage{
zero_hit(ll, ul)
}
\arguments{
\item{ll}{Lower limit.}

\item{ul}{Upper limit.}
}
\value{
Returns \code{TRUE} or \code{FALSE} based on the test.
}
\description{
Tests if the confidence intervals contain zero.
}
\seealso{
Other confidence interval miscellaneous functions: 
\code{\link{ci_hit}()},
\code{\link{ci_shape}()},
\code{\link{ci_width}()}
}
\concept{confidence interval miscellaneous functions}
\keyword{ci}
